+++ tests/testthat/test-trace.R
@@ -211,7 +211,8 @@ test_that("eval() frames are collapsed", {
 
 test_that("%>% frames are collapsed", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -235,7 +236,8 @@ test_that("%>% frames are collapsed", {
 
 test_that("children of collapsed %>% frames have correct parent", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -296,7 +298,8 @@ test_that("pipe_collect_calls() collects calls", {
 
 test_that("combinations of incomplete and leading pipes collapse properly", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -331,7 +334,8 @@ test_that("combinations of incomplete and leading pipes collapse properly", {
 
 test_that("calls before and after pipe are preserved", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")